requirements.txt
openpyxl~=3.1.2
selenium~=4.15.2
python-dotenv~=1.0.0